+++ crates/encrust-macros/tests/encrust_files.rs
@@ -12,7 +12,7 @@ fn encrust_file_string() {
     let mut encrust_file = encrust_macros::encrust_file_string!("tests/encrust_files.rs");
     let file = encrust_file.decrust();
 
-    assert_eq!(orig_file, file.as_str());
+    assert_eq!(orig_file, &*file);
 }
 
 #[test]
@@ -21,5 +21,5 @@ fn encrust_file_bytes() {
     let mut encrust_file = encrust_macros::encrust_file_bytes!("tests/encrust_files.rs");
     let file = encrust_file.decrust();
 
-    assert_eq!(orig_file.as_slice(), file.as_slice());
+    assert_eq!(orig_file.as_slice(), &*file);
 }