+++ CHANGELOG.md
@@ -3,6 +3,41 @@
 ## Changed
 * Changed the name of the `Encrustable` trait to `Encrust` to align the name better with usual
   naming of traits in Rust.
+* Made major changes to the `Encrust` trait.
+  * New types that must be defined:
+    * `Storage` - what type `Encrusted` should use for storing the type. This was introduced to
+      avoid undefined behavior with `String`s as they require that the data stored in the `String`s
+      are valid UTF-8 at all times, even when its not accessed.
+    * `Ref` - the type that `Decrusted` provides a reference to. This allows restricting access to
+      encrusted values. `String` defines `Ref = str` and `Vec<T>` defines `Ref = [T]`. This prevents
+      accidentally pushing new data to the values which could leave old, unobfuscated data in
+      memory.
+  * New functions that must be defined:
+    * `fn to_storage(self) -> Self::Storage` - to prepare a value for storage in `Encrusted`.
+    * `unsafe fn as_ref(storage: &Self::Storage) -> &Self::Ref` - convert from a reference to a
+      stored value to a reference to `Ref`. Calling `as_ref` on encrusted data may lead to
+      undefined behavior.
+    * `unsafe fn as_mut_ref(storage: &mut Self::Storage) -> &mut Self::Ref` - same as `as_ref`,
+      only with mutable references.
+  * `toggle_encrust` is no longer unsafe. It should always be safe to call `toggle_encrust`, but
+    accessing an encrusted value through `as_ref` or `as_mut_ref` may lead to undefined behavior if
+    the value is encrusted.
+  * `Encrust` is no longer implemented for arrays and `Vec`s of types that implement `Encrust`.
+    Instead, the arrays and `Vec`s must now be of types that implement the new trait
+    `InPlaceEncrust`.
+* MSRV changed from 1.85 to 1.87.
+
+## Added
+* Introduced a `InPlaceEncrust` trait for simple types such as integers where `Storage` and `Ref`
+  can be `Self`. `Encrust` is implemented for all `InPlaceEncrust` types.
+
+## Removed
+* Removed the derive macro for `Encrust` as a general derive macro does not work with the changed
+  `Encrust` trait.
+* Removed the `encrust_vec!` macro as encrusted `Vec`s no longer allow access to the underlying
+  `Vec`, but simply a slice to the `Vec`'s memory. With this change, `encrust_vec!([...])` is simply
+  a worse alternative to `encrust!([...])` as decrusting does not allow growing the `Vec`'s memory.
+* Removed `std` feature as it did not add anything.
 
 # Version 0.3.1 and 0.3.2 - 2025-11-27
 * Remove configuration and annotations that are no longer needed to generate documentation.